feat(authorization): identify administrative user by camunda-admin group
Administrative users for camunda BPM must be members of the group
`camunda-admin` (Groups.CAMUNDA_ADMIN). The camunda web
applications will check, whether a user in such a group exists and if
not will ask the user to create such an administrative user.

- update test data and setup resources accordingly

BREAKING CHANGE

Administrative users are now identified by membership in group
`camunda-admin`. Add a user with that group to a process engine
if you would like to prevent the initial setup dialog.

Closes to CAM-1029