feat(historic/incident): persist incidents as historic incidents
This commit adds the feature that incidents will be permanently stored in
the database, if the history level is equal to "FULL".

A HistoricIncident has all the applicalbe properties provided by runtime
IncidentEntity and in addition it provide also the following properties:

* createTime (equal to incident timestamp of IncidentEntity)
* endTime (represents the time when the incident has been resolved or
  deleted)
* incidentState:
  - open
  - deleted
  - resolved

Furthermore this commit adds a HistoricIncidentQuery to query after
HistoricIncidents.

related to CAM-1945