feat(cockpit/searchParams): keep search parameters on view change
This commit adds the possibility to define which search parameters should
be kept if the user change view for example from "Runtime" to "History"
(and vice versa). Therefore it is possible to set the option
"keepSearchParams" to true or to assign an array of search parameters which
should be kept.

related to CAM-2003