fix(variable/concurrent): add unique constraint
The commit adds a new column "VAR_SCOPE_" to the table "ACT_RU_VARIABLE". This new column is
used together with the column "NAME_" for an unique constraint. This
unique constraint prevents the possibilty to add the same variable to the same
execution (or task) in a concurrent way.

related to CAM-1574