feat(close): close a completed Case Instance
This commit adds the feature to close a completed case instance, so that
no further work or modifications are allowed for the associated case
instance (ie. the case instance has been removed from the runtime and is
not available anymore - note: at the moment no history is written for case
instances).

A case instance can be closed as follows:

  caseService
    .withCaseExecution("aCaseInstanceId")
    .close();

related to CAM-2197