chore(db-scripts): remove foreign key constraint on SUPER_CASE_EXEC_
This commit removes an unnecessary foreign key constraint on column
SUPER_CASE_EXEC_ inside the db table ACT_RU_CASE_EXECUTION.

If a CaseTask is non-blocking then the associated case execution will be
completed after the instantiation of the called case instance. During the
completion of the CaseTask the associated case execution will be removed
and with the foreign key constraint this would not be possible, because
the referential integrity would not be satisfied.

related to CAM-2444