feat(case/caseTask): execute a case task
This commit provides the implementation of the CaseTask.

If a case task becomes "ACTIVE" a given case will be instantiated,
e.g:

  <caseTask id="anId" caseRef="aCaseToInstantiate" ... />

In that case a new case instance will be started by the given case
definition key ("aCaseToInstantiate").

Furthermore it is possible to declare one of the following
"caseBinding":

- latest: means that a case instance of the latest version of the
  given case definition will be instantiated.
- deployment: means that a case instance of the case definition
  which is inside the same deployment as the case definition will be
  instantiated.
- version: means that a case instance of a declared
  "caseVersion" will be instantiated (therefore you need to provide
  a further attribute called "caseVersion").

Additionally it is possible to declare which variables should be exchanged
between the case instances (and vice versa), e.g.:

  <caseTask ... >
     <extensionElements>
       <camunda:in source="a" target="b" />
       <camunda:out source="b" target="c" />
     </extensionElements>
  </caseTask>

And the BusinessKey like this:

  <caseTask ... >
    <extensionElements>
      <camunda:in businessKey="${caseExecution.caseInstanceBusinessKey} />
    </extensionElements>
  </caseTask>

related to CAM-2444