fix(case/processTask): completion of sub instance
This commit fixes the completion behavior of a CaseTask and ProcessTask:
It is not possible to complete a CaseTask or Process task manually.

If the called Case instance completes the first time, this trigger
notification of the associated CaseTask and the associated CaseTask will
be completed automatically.

Furthermore termination or suspending a CaseTask or ProcessTask does not
have any effect on the called sub case or process instance (and vice
versa). Note: Termination and suspending is not part of the public API!

related to CAM-2371, CAM-2444