chore(milestone): provide test cases for milestones
This commit adds missing test cases for the implementation of the
milestone.

Note: Milestones does only make sense in combination with entryCriterias
(Sentries). This means if a milestone does not have any entryCriteria
this milestone will be completed immediately. But if a milestone
references at least one sentry as entryCriteria this milestone will not be
completed immediately. But in the current implementation it is not possible
to complete this milestone using the public Java API, because a milestone
will only completed if the entryCriterias are fulfilled (this behavior
will be implemented with the Sentries).

related to CAM-2385