fix(timer/job): schedule new job for next timer invocation
This commit "fixes" the behavior of a repeating timer job. Instead of
scheduling a new job for the next timer invocation after the repeating
timer job executes successfully, the new job will be scheduled when the
repeating timer job executes the first time (independet if the execution
succeeds or not). The repeating timer job saves that a new job has been
scheduled inside the job handler configuration field, e.g.:

  <activityId/definitionKey>$followUpJobCreated

This will also be persisted in case the repeating job fails.

related to CAM-2797