fix(engine/jobex): ExclusiveJobAddedNot. invoked from different Thread
PROBLEM: If the ExclusiveJobAddedNotification is invoked from another
thread than the job executor thread from which it was created, it
threw an NPE when attempting to access the JobExecutorContext from a
Thread Local.

FIX: Passing in the Job to the Listener fixes this problem.

NOTE: we cannot easily stage situations where this happens in
reality. This is why a unit test is added which makes sure there are
no regressions for this.

related to #CAM-3684