feat: add attribute errorCodeVariable to error event definition
In order to propagate an error payload to the catching execution the
error event definition now has an optional attribute "errorCodeVariable"
in the camunda: namespace. That ways the error code of the caught error
is usable as process variable. Especially after catch all error start or
boundary events the error code can be used to differentiate.

Related to CAM-2496