Add getter for AuthorizationService and CaseService
We started working on assertions for case plans and noticed that there are no getters (or setters) for the services in the TestRule. I made the rule implement ProcessEngineServices (so next time, this raises an error) and fixed the missing methods. Also, the setter for historyService was named setHistoricDataService ... I did not want to break the api, but it would be a good idea to make the old on deprecated ...