fix(engine): change createMembership to behave like save/create User
Change the createMembership to behave like save/crate Users and Groups.
The createMembership creates an MembershipEntity and hands it over to
the DbEntityManager so that it is saved in the cache and flushed with
the other Entities in the correct order.
The previous behavior is, that this method was bypassing the cache and
tried to write the Membership directly into the Database. That leads to
an Exception, when the User, Group and corresponding Membership are
written in one step (e.g. within the same Command Context)
My use case is a new identity service Plugin which delegates most
requests to the Camunda DBIdentitiyServiceProvider. During
authentication we want to update the Camunda DB.

related to CAM-4732