test(engine): Add test that writes multiple entities in one step
This Test case validates the fix, that fixes the problem of writing
multiple entities in one step into the database.
Therefore a Plugin that intercepts the checkPassword() method and writes
a User/Group and Membership into the database is necessary, to ensure
all of them are written in one step (the same Command Context). With the
new impl this test will be ok, with the old one it will fail.