fix(engine): make DmnDecisionTableResultMapping null save
If the decision doesn't match any rule (the result is empty) the
mapper singleEntry and singleRule now return null instead of an
exception. I prefer this solution as the user isn't able to
catch an exception on an empty result. With this approach
the user can check for null in the following process.

related to #CAM-4971