fix(engine/auth): perf of auth checks if user has many groups
improves performance of auth checks if user is part of many groups. This
fix has two items

1) use single case statement with GROUP_ID_ in (list of groups) instead
of generating many case statements, one per group the user is a member
of

2) once per command, check for which groups authorizations exist in the
db. Then interect that list with the list of groups the user is a member
of. This way, authorization checks are only performed for groups such
that the user is a member of the group and there exists at least one
authorization for that groups. We have seen cases where groups are
loaded from LDAP and the users were part of 200+ groups but only 1-2 of
these groups were actually used in Camunda.

related to #CAM-5364