Provide spring configuration for process engine services
I wrote this class 10 times so far, to expose all process services as beans in an annotation based environment. Not implementing ProcessEngineServices always lead to missing beans when upgrading the engine version (and new service where added). I would really like to have this on in the engine-spring core, so I can just use it in the future instead of c&p-ing it again.