fix(engine): UserOperationLogEntry not created for starting processes
Starting a process with a set of variables creates HistoricDetail entries.
These entries also carry a UserOperationId, however, that UserOperationId
can't be resolved to a UserOperationLogEntry.

Not being able to audit the creation of a process instance/the initial
setting of variables at process creation time is considered a bug.

This commit modifies the StartProcessInstance and 
StartProcessInstanceAtActivities to create the necessary 
UserOperationLogEntry.

This results in a larger set of UserOperationLogEntrys that will be generated
by the process engine, which needs to be accounted for/filtered accordingly.

Unittests that reproduce the problematic behavior were add, existing unittest
updated to work with the new situation.